+++ lib/page/assets/index.dart
@@ -387,8 +387,29 @@ class _AssetsState extends State<Assets> {
                   children: <Widget>[
                     Padding(
                       padding: EdgeInsets.only(top: 24),
-                      child: BorderedTitle(
-                        title: I18n.of(context).home['assets'],
+                      child: Row(
+                        crossAxisAlignment: CrossAxisAlignment.end,
+                        children: <Widget>[
+                          BorderedTitle(
+                            title: I18n.of(context).home['assets'],
+                          ),
+                          isAcala || isLaminar
+                              ? Expanded(
+                                  child: Padding(
+                                    padding:
+                                        EdgeInsets.only(bottom: 2, left: 8),
+                                    child: Text(
+                                      '(${I18n.of(context).assets['assets.test']})',
+                                      style: TextStyle(
+                                        fontSize: 14,
+                                        color: Theme.of(context)
+                                            .unselectedWidgetColor,
+                                      ),
+                                    ),
+                                  ),
+                                )
+                              : Container()
+                        ],
                       ),
                     ),
                     RoundedCard(